pyproject.toml
[tool.poetry]
name = "djrest-gitflow-int-mag"
version = "0.1.0"
description = ""
authors = ["Yaroslav Galkin <galkinyaroslav@gmail.com>"]
readme = "README.md"
packages = [{include = "djrest_gitflow_int_mag"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
djoser = "^2.1.0"
pytest-django = "^4.5.2"
prospector = "^1.9.0"
celery = "^5.2.7"
redis = "^4.5.4"
amqp = "^5.1.1"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
